DO NOT SUBMIT Use SHA-512 instead of Keccak in Kyber
The PRF and XOF is (roughly) MGF1-SHA512. The rest are plain SHA-512. In
each case, I've appended a single byte for domain separation. This is
also why it's only roughly MGF1-SHA512. I've put the domain separation
byte after the counter, instead of before, to ensure we have domain
separation between it and H, J, and G.

Put together for benchmarking.

Change-Id: Id8c9209ab7e95a715700abb652df23a6bde4b445