Remove a pile of unused X509_STORE callbacks
Other than the verify callback, these are practically unused. The one
exception is that gRPC have used a couple of the CRL functions, though
in somewhat questionable ways. Keep those for now, but we need to work
with them to fix their code.

There's also a bit of mess around check_issued, largely due to further
rust-openssl misunderstandings.

Update-Note: Removed a bunch of unused functions. We can restore them if
someone was using them.

Change-Id: I9c47581784c56a4c3b762e603a20ad7d5d612c65