Add a debugging log facility to the delegate.
This is to permit us to receive log messages with an integer
log level from the path builder, and we can then potentially
use them somewhere else (selectively based on the log level if
we so choose) to put the messages somewhere BoringSSL has no
knowledge of. In chrome this will enable us to send path builder
diagnostics to netlog.

Change-Id: I5c2e04a25db1e41c1b825f68ec68e4e587aa2994