Fix the names of some X509_STORE_CTX functions
This matches an upstream change. Add macros for the old names. We may be
able to unexport these, but for now just pick up the less confusing
names.

(I found only one user of one of them, goma, which will be replaced next
year. Though wanting *some* API to query the dirhash machinery is not
completely implausible.)

Change-Id: Idd2352c07c294c4f63c4ef12e5d97804f42225b9