Remove fillins/openssl_util
The only user of the tiny utility class to clear the error stack
was verify_certificate_chain, which can have it in it's anonymous
namespace.

It was however, used to great effect as a horrible dirty cheat to get
us openssl/base.h during the conversion. That time is past, so just
let us include <openssl/base.h> normally, and get rid of this.

Bug: 668
Change-Id: Ie8537d5b1d6789acb35182f3d781f7505ad79109