Consistently call CRYPTO_free_ex_data first
(Or second, if there's a method table.)

CRYPTO_EX_free is passed the parent object, so the caller could, in
principle, inspect the object. We should pass in an object in a
self-consistent state. Also fix up the documentation. I think some bits
of a since removed CRYPTO_EX_new got jumbled up in there.

Change-Id: I316d00aee61bf544f59d4dac2efcd825e4bfa9b1