Remove dynamic X509_TRUST and X509_PURPOSE registration
This is not thread-safe. Even if made thread-safe, it involves
registering globals, so it's just not a good API.

Note this means that there is no longer a way to configure custom trust
OIDs or purpose checks. Evidently no one was doing that. Should a use
case arise, I don't think it should be met by this API. The things one
might want to configure here are:

- Which OID to match against X509_add1_trust_object and
  X509_add1_reject_object

- Whether self-signed certificates, if no trust objects are configured,
  also count as trust anchors

- Which EKU OID to look for up the chain

- Which legacy Netscape certificate type to look for (can we remove
  this?)

- Which key usage bits to look for in the leaf

We can simply add APIs for specifying those if we need them.

Interestingly, there's a call to check_ca inside the purpose checks
(which gets skipped if you don't configure a purpose!), but I think it
may be redundant with the X509_check_ca call in the path verifier.

Change-Id: If71ee3d0768b5fc71422852b4fcf7eb23e937dd2