Assume the Arm assembler can handle ADR
It's 2023. We shouldn't need to be counting offsets from PC anymore.
Instead, let the assembler figure this out with an ADR instruction.
This reduces our dependencies on the exact offset between code and data
in the Arm assembly.

It only reduces it because we're still subject to limitations on the ADR
offset. 32-bit Arm has a hard time loading constants. All this logic
assumes the ADR point is very close to the assembly. For instance, an
ADR of K256 inside sha256_block_data_order_armv8 does not work because
K256 is too far away.

This seems to build fine with the one project I'm aware of that targets
Armv6.

Update-Note: If 32-bit Arm assembly no longer builds, let us know and
tell us what your toolchain is.

Cq-Include-Trybots: luci.boringssl.try:android_arm
Change-Id: Ie191781fed98d53c3b986b2f535132b970d79f98