Const-correct X509_alias_get0 and X509_keyid_get0
All callers I found seem to be compatible with this. Using the non-const
pointer isn't very useful because you cannot reise the value. Let's try
const-correcting it and we'll revert if it's too annoying to fix.

Update-Note: The above functions are now const-correct. Store the result
in a const pointer to avoid compatibility issues.

Change-Id: Id4a1c7223fbb333716906e20844bf8795118a8ea