Document and test X509_PURPOSE and X509_TRUST machinery
The trust and purpose is all a bit tied up together, as is the meaning
of the certificates in an X509_STORE at all. (It's hard to discuss
whether a "trusted certificate" is actually a trust anchor without a
description of trust settings to reference.)

Cut the Gordian Knot by documenting all that first. Later CLs will move
other symbols into the sections established here. Also as the behavior
is a little complex, add some tests to cover some of this machinery.

Bug: 426
Change-Id: Idde8bc4e588de92ebabf6ecf640b62a2a6803688