Reworking bssl_crypto: ECDH
This change also adds scoped.rs, which contains scoped objects for
`EVP_PKEY` and `EC_KEY`, for when we're holding temporary objects of
those types.

Due to an accident with git rebase, it also renames `crypto_memcmp` to
`constant_time_compare` and promotes it to the top-level of the crate.

Change-Id: I629c051a244e3f9dcf64d8a36846528f10a31f50