Do not condition CRYPTO_is_RDRAND_capable on __RDRND__
On some AMD processors, we will act as if the hardware is
RDRAND-incapable, even it actually supports it, to workaround some bugs
where their RDRAND would return all ones. However, this works by
clearing the RDRAND bit, and the static capability check bypasses this.

That means, when compiling binaries for an RDRAND-capable baseline CPU,
we might end up disabling the workaround. Thanks to Brian Smith for
noticing this.

Fortunately, we believe this had no real impact and the workaround was
actually redundant with other parts of our PRNG logic. See
https://crbug.com/boringssl/675#c2 for a more detailed analysis.

Fixed: 675
Change-Id: I91d844cfaafb22eb29254f81fb6719588937a4a8