Use four-iterator std::equal for bssl::Span::operator==
At one point we targetted C++11 and had to use the unsafe three-iterator
std::equal. Then we had to write the loop by hand because the pragma to
suppress an MSVC warning did not work:
https://boringssl-review.googlesource.com/c/boringssl/+/18384

We now target C++14, so just use the four-iterator version. libc++
will then be able to specialize it into a memcmp call.

Also replace a three-iterator std::equal in pki/input.cc.

Change-Id: I0a7d6a2330594e383c5c884632de9a21a137054c