Import tests for CVE-2024-0727
BoringSSL is not affected by CVE-2024-0727, but these are good cases to
have in our unit tests.

PKCS#12 is built on top of PKCS#7, a misdesigned, overgeneralized combinator
format. One of the features of PKCS#7 is that the content of every
ContentInfo may be omitted, to indicate that the value is "supplied by other
means". This is commonly used for "detached signatures", where the signature
is supplied separately.

This does not make sense in the context of PKCS#12. But because PKCS#7
combined many unrelated use cases into the same format, so PKCS#12 (and any
other use of PKCS#7) must account for and reject inputs.

Change-Id: I22f19b6c14894003f7515206cd34f968e5503d4a