Expose OPENSSL_timegm in posix_time.h
Ok so this one is purely for keeping this the same between forks.
Boring today doesn't need this (although some of our outside
uses might be able to use it). Effectively, this function
is the same as converting to a posix time, followed by a safe
check to see if you can put the result in a time_t.

posix_time.h is about to get added as public in LibreSSL, and
while not strictly necessary there because they could inline
what it does, Libre is finding that may be needed in rpki-client,
ocsp-client, as well as libtls (including for libtls portable).

  In the interests of keeping the same API in the same file, Libre
would like this to be here so it's just consistent from both places.

Change-Id: I2f234005e83cdea5e61fa41fbf03ef61516c45f8